backend/requirements.txt
langchain-groq
langchain-mcp-adapters
mcp
langgraph
langchain-google-genai
langchain
langchain-core
langchain-ollama
langchain-openai
google_auth_oauthlib
google-api-python-client
fastmcp
openai
python-dotenv
Pillow
requests
fastapi
uvicorn
pymongo
# Lead Generation dependencies
aiohttp
rapidfuzz
beautifulsoup4
trafilatura
playwright
google-search-results
newsapi-python